64-bit issue
git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@1082 a2d82226-602c-0410-bba7-e51b83dd10d2

+++ owamp/capi.c
@@ -628,7 +628,7 @@ _OWPClientRequestTestReadResponse(
     uint16_t        port_ret=0;
     uint8_t         *sid_ret=NULL;
     char            nodename_buf[255];
-    int             nodename_buflen;
+    size_t          nodename_buflen;
 
     if( (rc = _OWPWriteTestRequest(cntrl,retn_on_intr,
                     I2AddrSAddr(sender,NULL),
